Centralized logic on the isRewriting method.
The method isn't working correctly yet but we we're testing if
the url was being redirected by proxy or ModRewrite that would be
great to get that decided in one place.

Also added the strEndsWith as that is being used in a lot of
places.

Fixed some errors and warnings.

Refs: #3