Created a global handler to hold the output without send headers.
Headers can be set during the processing becuase the content only
will be flushed only at the end.

This will make us to set correctly the content type using the
framework.

The way we were handling the content type was bad because the
header was set at the beginning of the parsed code and originating
the bug number #11.

This commit will make possible to set content type and charset
without problems and set headers without flushing output garanteed.

Refs: #11