Introduced security parameter into application, circuit and action.
The default security mode for application is optimistic. If the
parameter was set invalid, optimistic is used.

Circuit will get default from the security value from it's
appplication if the parameter isn't defined or invalid.

Applcaiont will get default from the security value from it's
circuit if the parameter isn't defined or invalid.

Refs: #85