Getting protocol from the request sheme on the server global array.
This is being implemented by the new getProtocol method. If request
scheme isn't present, will return http by default.

We need to add a parameter to the application to force the
resolution of the protocol via configuration.

Also some refactoring as made on this commit.

Tests for getProtocol and getRootUrl were added to the projec.

Refs: #1, #3